backend/requirements.txt
aiosqlite==0.22.1
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.31.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.2
diskcache==5.6.3
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.136.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
llama_cpp_python==0.3.20
Mako==1.3.11
MarkupSafe==3.0.3
numpy==2.4.4
packaging==26.1
passlib==1.7.4
pluggy==1.6.0
pycparser==3.0
pydantic==2.13.2
pydantic-settings==2.13.1
pydantic_core==2.46.2
Pygments==2.20.0
PyJWT==2.12.1
pytest==9.0.3
pytest-asyncio==1.3.0
python-dotenv==1.2.2
python-multipart==0.0.26
redis==7.4.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
